refactor(frontend): ControlsSettings を CharacterSettings に改名
- Replaced the ControlsSettings component with CharacterSettings in relevant files.
- Updated import paths and references to reflect the new component name.
- Adjusted the SETTINGS_TAB_LABELS to change 'Controls' to 'Character' for consistency.
