+++ frontend/src/app/components/SettingsPanel.tsx
@@ -6,7 +6,7 @@ import {
   MessageSquare,
   Palette,
   Presentation,
-  SlidersHorizontal,
+  User,
   Volume2,
   X,
 } from 'lucide-react';
@@ -16,10 +16,10 @@ import AudioSettings from './AudioSettings';
 import BackgroundSelector, {
   type BackgroundOption as BackgroundSelectorOption,
 } from './BackgroundSelector';
-import ControlsSettings, {
+import CharacterSettings, {
   type ControlsState,
   type VRMAnimationOption,
-} from './ControlsSettings';
+} from './CharacterSettings';
 import EnvironmentSettings from './EnvironmentSettings';
 import KeyframeSettings from './KeyframeSettings';
 
@@ -27,7 +27,7 @@ const SETTINGS_TAB_LABELS: Record<
   'controls' | 'keyframe' | 'background' | 'lighting' | 'audio' | 'slides',
   string
 > = {
-  controls: 'Controls',
+  controls: 'Character',
   keyframe: 'Keyframe',
   background: 'Background',
   lighting: 'Lighting',
@@ -48,7 +48,7 @@ export interface SettingsPanelProps {
   show_close_button?: boolean;
   on_close?: () => void;
   extra_tab?: { label: string; content: ReactNode };
-  /** Controls tab (uses ControlsSettings component) */
+  /** Character tab (uses CharacterSettings component) */
   controls_state: ControlsState;
   vrm_expression_names: string[];
   expression_weights: Record<string, number>;
@@ -170,7 +170,7 @@ export default function SettingsPanel({
           >
             {tab === 'conversation' && <MessageSquare className="size-4" />}
             {tab === 'slides' && <Presentation className="size-4" />}
-            {tab === 'controls' && <SlidersHorizontal className="size-4" />}
+            {tab === 'controls' && <User className="size-4" />}
             {tab === 'keyframe' && <Film className="size-4" />}
             {tab === 'background' && <Palette className="size-4" />}
             {tab === 'lighting' && <Lightbulb className="size-4" />}
@@ -212,7 +212,7 @@ export default function SettingsPanel({
 
       {active_tab === 'controls' ? (
         <div className="mb-4">
-          <ControlsSettings
+          <CharacterSettings
             state={controls_state}
             vrmExpressionNames={vrm_expression_names}
             expressionWeights={expression_weights}
